pipeline-rag-mvp/requirements.txt
python-docx>=1.1.0
PyYAML>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
